Add itii in models + separate files
